+++ crates/pop-cli/tests/contract.rs
@@ -35,7 +35,7 @@ async fn contract_lifecycle() -> Result<()> {
 		.args(&["build", "--path", "./test_contract", "--release"])
 		.assert()
 		.success();
-	// Verify that the folder target has been created
+	// Verify that the directory target has been created
 	assert!(temp_dir.join("test_contract/target").exists());
 	// Verify that all the artifacts has been generated
 	assert!(temp_dir.join("test_contract/target/ink/test_contract.contract").exists());
@@ -71,7 +71,8 @@ async fn contract_lifecycle() -> Result<()> {
 		])
 		.assert()
 		.success();
-	// Using methods from the pop_contracts crate to instantiate it to get the Contract Address for the call
+	// Using methods from the pop_contracts crate to instantiate it to get the Contract Address for
+	// the call
 	let instantiate_exec = set_up_deployment(UpOpts {
 		path: Some(temp_dir.join("test_contract")),
 		constructor: "new".to_string(),