+++ crates/pop-parachains/src/build.rs
@@ -14,10 +14,12 @@ use std::{
 /// Build the parachain and returns the path to the binary.
 ///
 /// # Arguments
-/// * `path` - The optional path to the parachain manifest, defaulting to the current directory if not specified.
+/// * `path` - The optional path to the parachain manifest, defaulting to the current directory if
+///   not specified.
 /// * `package` - The optional package to be built.
 /// * `release` - Whether the parachain should be built without any debugging functionality.
-/// * `node_path` - An optional path to the node directory. Defaults to the `node` subdirectory of the project path if not provided.
+/// * `node_path` - An optional path to the node directory. Defaults to the `node` subdirectory of
+///   the project path if not provided.
 pub fn build_parachain(
 	path: &Path,
 	package: Option<String>,
@@ -33,13 +35,14 @@ pub fn build_parachain(
 		args.push("--release");
 	}
 	cmd("cargo", args).dir(path).run()?;
-	binary_path(&profile.target_folder(path), node_path.unwrap_or(&path.join("node")))
+	binary_path(&profile.target_directory(path), node_path.unwrap_or(&path.join("node")))
 }
 
 /// Determines whether the manifest at the supplied path is a supported parachain project.
 ///
 /// # Arguments
-/// * `path` - The optional path to the manifest, defaulting to the current directory if not specified.
+/// * `path` - The optional path to the manifest, defaulting to the current directory if not
+///   specified.
 pub fn is_supported(path: Option<&Path>) -> Result<bool, Error> {
 	let manifest = from_path(path)?;
 	// Simply check for a parachain dependency
@@ -51,7 +54,7 @@ pub fn is_supported(path: Option<&Path>) -> Result<bool, Error> {
 	}))
 }
 
-/// Constructs the node binary path based on the target path and the node folder path.
+/// Constructs the node binary path based on the target path and the node directory path.
 ///
 /// # Arguments
 /// * `target_path` - The path where the binaries are expected to be found.
@@ -121,7 +124,8 @@ pub fn generate_raw_chain_spec(
 /// Export the WebAssembly runtime for the parachain.
 ///
 /// # Arguments
-/// * `binary_path` - The path to the node binary executable that contains the `export-genesis-wasm` command.
+/// * `binary_path` - The path to the node binary executable that contains the `export-genesis-wasm`
+///   command.
 /// * `chain_spec` - Location of the raw chain specification file.
 /// * `wasm_file_name` - The name of the wasm runtime file to be generated.
 pub fn export_wasm_file(
@@ -152,7 +156,8 @@ pub fn export_wasm_file(
 /// Generate the parachain genesis state.
 ///
 /// # Arguments
-/// * `binary_path` - The path to the node binary executable that contains the `export-genesis-state` command.
+/// * `binary_path` - The path to the node binary executable that contains the
+///   `export-genesis-state` command.
 /// * `chain_spec` - Location of the raw chain specification file.
 /// * `genesis_file_name` - The name of the genesis state file to be generated.
 pub fn generate_genesis_state_file(
@@ -348,7 +353,7 @@ mod tests {
 		Ok(())
 	}
 
-	// Function that generates a Cargo.toml inside node folder for testing.
+	// Function that generates a Cargo.toml inside node directory for testing.
 	fn generate_mock_node(temp_dir: &Path) -> Result<(), Error> {
 		// Create a node directory
 		let target_dir = temp_dir.join("node");
@@ -414,12 +419,12 @@ default_command = "pop-node"
 		cmd("cargo", ["new", name, "--bin"]).dir(temp_dir.path()).run()?;
 		generate_mock_node(&temp_dir.path().join(name))?;
 		let binary = build_parachain(&temp_dir.path().join(name), None, &Profile::Release, None)?;
-		let target_folder = temp_dir.path().join(name).join("target/release");
-		assert!(target_folder.exists());
-		assert!(target_folder.join("parachain_template_node").exists());
+		let target_directory = temp_dir.path().join(name).join("target/release");
+		assert!(target_directory.exists());
+		assert!(target_directory.join("parachain_template_node").exists());
 		assert_eq!(
 			binary.display().to_string(),
-			target_folder.join("parachain_template_node").display().to_string()
+			target_directory.join("parachain_template_node").display().to_string()
 		);
 		Ok(())
 	}